Group utility operations into same file.
