+++ pennylane/ops/qubit/non_parametric_ops.py
@@ -2334,132 +2334,3 @@ class MultiControlledX(Operation):
     @property
     def is_hermitian(self):
         return True
-
-
-class Barrier(Operation):
-    r"""Barrier(wires)
-    The Barrier operator, used to separate the compilation process into blocks or as a visual tool.
-
-    **Details:**
-
-    * Number of wires: AnyWires
-    * Number of parameters: 0
-
-    Args:
-        only_visual (bool): True if we do not want it to have an impact on the compilation process. Default is False.
-        wires (Sequence[int] or int): the wires the operation acts on
-    """
-    num_params = 0
-    """int: Number of trainable parameters that the operator depends on."""
-
-    num_wires = AnyWires
-    par_domain = None
-
-    def __init__(self, wires=Wires([]), only_visual=False, id=None):
-        self.only_visual = only_visual
-        self.hyperparameters["only_visual"] = only_visual
-        super().__init__(wires=wires, id=id)
-
-    @staticmethod
-    def compute_decomposition(wires, only_visual=False):  # pylint: disable=unused-argument
-        r"""Representation of the operator as a product of other operators (static method).
-
-        .. math:: O = O_1 O_2 \dots O_n.
-
-
-        .. seealso:: :meth:`~.Barrier.decomposition`.
-
-        ``Barrier`` decomposes into an empty list for all arguments.
-
-        Args:
-            wires (Iterable, Wires): wires that the operator acts on
-            only_visual (Bool): True if we do not want it to have an impact on the compilation process. Default is False.
-
-        Returns:
-            list: decomposition of the operator
-
-        **Example:**
-
-        >>> print(qml.Barrier.compute_decomposition(0))
-        []
-
-        """
-        return []
-
-    def label(self, decimals=None, base_label=None, cache=None):
-        return "||"
-
-    def _controlled(self, _):
-        return copy(self).queue()
-
-    def adjoint(self):
-        return copy(self)
-
-    def pow(self, z):
-        return [copy(self)]
-
-    def simplify(self):
-        if self.only_visual:
-            if len(self.wires) == 1:
-                return qml.Identity(self.wires[0])
-            return qml.prod(*(qml.Identity(w) for w in self.wires))
-        return self
-
-
-class WireCut(Operation):
-    r"""WireCut(wires)
-    The wire cut operation, used to manually mark locations for wire cuts.
-
-    .. note::
-
-        This operation is designed for use as part of the circuit cutting workflow.
-        Check out the :func:`qml.cut_circuit() <pennylane.cut_circuit>` transform for more details.
-
-    **Details:**
-
-    * Number of wires: AnyWires
-    * Number of parameters: 0
-
-    Args:
-        wires (Sequence[int] or int): the wires the operation acts on
-    """
-    num_params = 0
-    num_wires = AnyWires
-    grad_method = None
-
-    def __init__(self, *params, wires=None, id=None):
-        if wires == []:
-            raise ValueError(
-                f"{self.__class__.__name__}: wrong number of wires. "
-                f"At least one wire has to be given."
-            )
-        super().__init__(*params, wires=wires, id=id)
-
-    @staticmethod
-    def compute_decomposition(wires):  # pylint: disable=unused-argument
-        r"""Representation of the operator as a product of other operators (static method).
-
-        Since this operator is a placeholder inside a circuit, it decomposes into an empty list.
-
-        Args:
-            wires (Any, Wires): Wire that the operator acts on.
-
-        Returns:
-            list[Operator]: decomposition of the operator
-
-        **Example:**
-
-        >>> print(qml.WireCut.compute_decomposition(0))
-        []
-
-        """
-        return []
-
-    def label(self, decimals=None, base_label=None, cache=None):
-        return "//"
-
-    def adjoint(self):
-        return WireCut(wires=self.wires)
-
-    def pow(self, z):
-        return [copy(self)]
